iarm-azure-service/requirements.txt
azure-cognitiveservices-speech==1.43.0
python-dotenv==1.1.0
openai==1.75.0